Black/white mode and new text based image class
Finished the implementation of black/white mode, instructions have been added to the readme. A new class named "Label" has been added to allow for better display of text on the screen. Also added was a title area above the gameboard, this will be later used to display all messages to the user.